+++ nequip/nn/_tp_scatter_oeq.py
@@ -1,10 +1,4 @@
-import torch
 from ._tp_scatter_base import TensorProductScatter
-from openequivariance import (
-    TensorProductConv,
-    TPProblem,
-    torch_to_oeq_dtype,
-)
 
 
 class OpenEquivarianceTensorProductScatter(TensorProductScatter):
@@ -15,6 +9,7 @@ class OpenEquivarianceTensorProductScatter(TensorProductScatter):
         irreps_edge_attr,
         irreps_mid,
         instructions,
+        use_opaque: bool,
     ) -> None:
         super().__init__(
             feature_irreps_in=feature_irreps_in,
@@ -25,20 +20,37 @@ class OpenEquivarianceTensorProductScatter(TensorProductScatter):
         # ^ we ensure that the base class keeps around a `self.tp` that carries its own set of persistent buffers
         # even though `self.tp` is not used, having its (persistent) buffers always around ensures state dict compatibility when adding on or removing this subclass module
 
-        default_dtype = torch.get_default_dtype()
+        # === OEQ ===
 
-        # OEQ
+        # we do lazy imports of oeq to allow `nequip-package` to pick this file up even if oeq is not installed
+        # since `nequip-package` ignores files if it errors on loading the file
+
+        from openequivariance import (
+            TensorProductConv,
+            TPProblem,
+            torch_to_oeq_dtype,
+        )
+
         tpp = TPProblem(
             feature_irreps_in,
             irreps_edge_attr,
             irreps_mid,
             instructions,
-            irrep_dtype=torch_to_oeq_dtype(default_dtype),
-            weight_dtype=torch_to_oeq_dtype(default_dtype),
+            irrep_dtype=torch_to_oeq_dtype(self.model_dtype),
+            weight_dtype=torch_to_oeq_dtype(self.model_dtype),
             shared_weights=False,
             internal_weights=False,
         )
-        self.tp_conv = TensorProductConv(tpp, torch_op=True, deterministic=False)
+        self.tp_conv = TensorProductConv(
+            tpp, torch_op=True, deterministic=False, use_opaque=use_opaque
+        )
 
     def forward(self, x, edge_attr, edge_weight, edge_dst, edge_src):
-        return self.tp_conv(x, edge_attr, edge_weight, edge_dst, edge_src)
+        # explicit cast to account for AMP
+        return self.tp_conv(
+            x.to(self.model_dtype),
+            edge_attr.to(self.model_dtype),
+            edge_weight.to(self.model_dtype),
+            edge_dst,
+            edge_src,
+        )