+++ nequip/nn/graph_model.py
@@ -22,7 +22,7 @@ def _model_metadata_from_config(model_config: Dict[str, str]) -> Dict[str, str]:
     model_metadata_dict[NUM_TYPES_KEY] = str(len(model_config[TYPE_NAMES_KEY]))
     model_metadata_dict[R_MAX_KEY] = str(model_config[R_MAX_KEY])
 
-    if PER_EDGE_TYPE_CUTOFF_KEY in model_config:
+    if model_config.get(PER_EDGE_TYPE_CUTOFF_KEY, None) is not None:
         from .embedding.utils import per_edge_type_cutoff_to_metadata_str
 
         model_metadata_dict[PER_EDGE_TYPE_CUTOFF_KEY] = (
@@ -47,6 +47,7 @@ class GraphModel(GraphModuleMixin, torch.nn.Module):
 
     model_input_fields: List[str]
     is_graph_model: Final[bool] = True
+    is_compile_graph_model: Final[bool] = False
     # ^ to identify `GraphModel` types from `nequip-package`d models (see https://pytorch.org/docs/stable/package.html#torch-package-sharp-edges)
 
     _metadata: Dict[str, str]