+++ nequip/scripts/package.py
@@ -1,13 +1,5 @@
 # This file is a part of the `nequip` package. Please see LICENSE and README at the root for information on using it.
-import argparse
 
-import pathlib
-import yaml
-
-# TODO: check if we still need this?
-# This is a weird hack to avoid Intel MKL issues on the cluster when this is called as a subprocess of a process that has itself initialized PyTorch.
-# Since numpy gets imported later anyway for dataset stuff, this shouldn't affect performance.
-import numpy as np  # noqa: F401
 import torch
 
 from nequip.model.saved_models.checkpoint import data_dict_from_checkpoint
@@ -17,11 +9,6 @@ from nequip.model.saved_models.package import (
     _suppress_package_importer_warnings,
 )
 from nequip.model.saved_models import ModelFromCheckpoint
-from nequip.model.saved_models.package import (
-    _EXTERNAL_MODULES,
-    _MOCK_MODULES,
-    _INTERNAL_MODULES,
-)
 from nequip.model.utils import (
     _COMPILE_MODE_OPTIONS,
     _EAGER_MODEL_KEY,
@@ -30,13 +17,22 @@ from nequip.nn.model_modifier_utils import is_persistent_model_modifier
 from nequip.model.modify_utils import get_all_modifiers, only_apply_persistent_modifiers
 from nequip.utils.logger import RankedLogger
 from nequip.utils.versions import get_current_code_versions, _TORCH_GE_2_6
-from nequip.utils.version_utils import get_version_safe
+from nequip.utils.versions.version_utils import get_version_safe
 from nequip.utils.global_state import set_global_state
 
 from ._workflow_utils import set_workflow_state
+from ._package_utils import (
+    _EXTERNAL_MODULES,
+    _MOCK_MODULES,
+    _INTERNAL_MODULES,
+)
 
 from omegaconf import OmegaConf
 import hydra
+import argparse
+import pathlib
+import yaml
+
 
 # === setup logging ===
 hydra.core.utils.configure_log(None)